requirements.txt
pyyaml
netaddr